+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [next]
 
+- BREAKING refactor(Intersection): support ray intersection [#9371](https://github.com/fabricjs/fabric.js/pull/9371)
+  **BREAKING**
+  - `Intersection.intersectsLineLine(a1, a2, b1, b2, boolean, boolean)` => `Intersection.intersectsLineLine(a1, a2, b1, b2, LineType, LineType)`
 - fix(Object): fixes centeredScaling prop type [#9401](https://github.com/fabricjs/fabric.js/pull/9401)
 - CI(): fix build caching + tests when merging to master [#9404](https://github.com/fabricjs/fabric.js/pull/9404)
 - chore(): export poly control utils [#9400](https://github.com/fabricjs/fabric.js/pull/9400)